Extract duplicated test helpers into shared scope
